Move logging code into a separate file.
